explicit casts to make xcode happy

+++ filter/fast_detector/fast_9.cpp
@@ -2970,7 +2970,7 @@ vector<xy> &fast_detector::detect(const unsigned char *im, const unsigned char *
 		end_if:
 
                         if(bmin == bmax - 1 || bmin == bmax) {
-                            features.push_back((xy){x, y, bmin, 0});
+                            features.push_back((xy){(float)x, (float)y, (float)bmin, 0});
                             push_heap(features.begin(), features.end(), xy_comp);
                             if(features.size() > need) {
                                 pop_heap(features.begin(), features.end(), xy_comp);
